feat(macro): a wolf can stand in a roster - one id space for men and beasts
A squad is a squad whatever it is made of (CANON.md S4/S16), but the code
could not say so. SoldierRecord.kind was a BYTE holding an NPCType, while the
monster half of the one id space every body already shares with the ECS is
0x100 | catalog row - it does not fit. Two consequences, both silent: a beast
could never be a squad member, and write_macro_npc refused outright any macro
entity whose kind was not humanoid, so a pack could not exist on the map at
all. The deserter pool inherited the same hole: kill a wolfpack's leader and
its survivors could not even fall in.

The field is now 16 bits, in SoldierRecord and in ecs::SoldierLink, and
valid_npc_kind answers for both halves (below 0x100 a row of kNpcTypeDefs, at
or above it one of the monster catalog). The macro snapshot's guard changed
meaning with it: it used to refuse monsters, it now refuses only a kind that
names NO row - a corrupt entity rather than a beast.

Where a member becomes a body, the row decides which birth it gets: a
humanoid ordinal builds a sheet-bearing body and takes the leader's aura, a
monster row builds a sheet-less one off its own catalog line. Both carry the
same roster receipt, so a wolf's death pays the pack back exactly like a
spearman's.

Upkeep answers honestly for the new half: a beast draws no pay, because upkeep
is a wage and nobody wages a wolf. Its cost to a squad is what it eats out of
the world, not out of a purse.

kSaveVersion 41 -> 42: the record's layout changed.

Two negative controls, both run and both seen failing before the fix: close
the validity gate to monsters again and save_roundtrip_test's beast member
stops surviving the round trip; disable the body fork and
subworld_spawn_parity_test's new beast-member case reports the pack arriving
as men. 76/76 ctest green, zero warnings.

Still humanoid-only, and deliberately out of this increment: a monster as the
macro squad's LEADER. make_npc builds a lord (sheet, traits, name, loot bag)
and a beast needs its own birth there - that is the prerequisite for a Lair
raising its own agents, which is the next step.

+++ src/ecs/components.h
@@ -126,7 +126,11 @@ struct NpcTraits {
 // roster remains authoritative; the ECS entity is a session projection.
 struct SoldierLink {
     std::uint32_t entityId;
-    std::uint8_t  kind;
+    // The same 16-bit kind a roster record carries (macro/army.h): a humanoid
+    // ordinal below 0x100, a monster catalog row at or above it. A beast in a
+    // squad is an ordinary member, so its link has to be able to say which
+    // beast.
+    std::uint16_t kind;
     std::int16_t  level;
 };
